refactor: switch containers to mimalloc-aware aliases and use sel_t for row indices (#546)
This pull request introduces a new container abstraction to standardize
and optimize memory allocation for column data structures, with support
for mimalloc-based allocation when enabled. It replaces standard STL
containers with custom types throughout the codebase, updates function
signatures to use the new types, and ensures consistent usage in edge
and context column classes. Additionally, it adds a compile-time flag
for mimalloc integration.

Container abstraction and allocator support:

* Added `container_types.h` defining `neug::vector`,
`neug::flat_hash_map`, `neug::flat_hash_set`, and allocator selection
based on the `NEUG_WITH_MIMALLOC` macro. Also introduced `sel_t` and
`sel_vec_t` as selection index types.
* Updated `CMakeLists.txt` to add the `NEUG_WITH_MIMALLOC` compile
definition and changed mimalloc override settings for more controlled
allocator usage.


These changes improve memory management flexibility, enable allocator
customization, and standardize container usage across the codebase.

Fixes

---------

+++ include/neug/execution/common/operators/retrieve/dedup.h
@@ -26,7 +26,7 @@ namespace execution {
 class Dedup {
  public:
   static neug::result<ContextChunk> dedup(ContextChunk&& chunk,
-                                          const std::vector<size_t>& cols);
+                                          const std::vector<int32_t>& cols);
 };
 
 }  // namespace execution
